Adding new app role for revoke certificate
