+++ KeyVault.Acmebot/Functions/GetCertificates.cs
@@ -6,7 +6,6 @@ using Azure.WebJobs.Extensions.HttpApi;
 
 using DurableTask.TypedProxy;
 
-using KeyVault.Acmebot.Internal;
 using KeyVault.Acmebot.Models;
 
 using Microsoft.AspNetCore.Http;
@@ -39,7 +38,7 @@ public class GetCertificates : HttpFunctionBase
         [DurableClient] IDurableClient starter,
         ILogger log)
     {
-        if (!User.IsAppAuthorized())
+        if (!User.Identity.IsAuthenticated)
         {
             return Unauthorized();
         }
